fix: Namespaced obstacle
