+++ app/graphql/types/tfl/entities/journey_planner/obstacle_type.rb
@@ -1,0 +1,14 @@
+module Types
+  module Tfl
+    module Entities
+      module JourneyPlanner
+        class ObstacleType < Types::BaseObject
+          field :incline, String, null: true
+          field :position, String, null: true
+          field :stopId, Integer, null: true
+          field :type, String, null: true
+        end
+      end
+    end
+  end
+end
